+++ src/App/Service/ClassGenerator.php
@@ -14,6 +14,7 @@ namespace App\Service;
 use App\Entity\Service;
 use Symfony\Bundle\MakerBundle\FileManager;
 use Symfony\Bundle\MakerBundle\Generator;
+use \Exception;
 
 /**
  * @author Christian Siewert <christian@sieware.international>
@@ -54,7 +55,7 @@ class ClassGenerator
     /**
      * @param Service $service
      * @return string
-     * @throws \Exception
+     * @throws Exception
      */
     public function generateEntityClass(Service $service) : string
     {
@@ -82,7 +83,7 @@ class ClassGenerator
     /**
      * @param Service $service
      * @return string
-     * @throws \Exception
+     * @throws Exception
      */
     public function generateRepositoryClass(Service $service) : string
     {
@@ -114,5 +115,53 @@ class ClassGenerator
     public function getTemplatePath(string $template = 'Entity.tpl.php') : string
     {
         return dirname(__DIR__) . sprintf(self::TEMPLATE_PATH, $template);
+    }
+
+    /**
+     * @param string $annotationClass The annotation: e.g. "@ORM\Column"
+     * @param array  $options         Key-value pair of options for the annotation
+     *
+     * @return string
+     */
+    public function buildAnnotationLine(string $annotationClass, array $options)
+    {
+        $formattedOptions = array_map(function ($option, $value) {
+            if (is_array($value)) {
+                if (!isset($value[0])) {
+                    return sprintf('%s={%s}', $option, implode(', ', array_map(function ($val, $key) {
+                        return sprintf('"%s" = %s', $key, $this->quoteAnnotationValue($val));
+                    }, $value, array_keys($value))));
+                }
+
+                return sprintf('%s={%s}', $option, implode(', ', array_map(function ($val) {
+                    return $this->quoteAnnotationValue($val);
+                }, $value)));
+            }
+
+            return sprintf('%s=%s', $option, $this->quoteAnnotationValue($value));
+        }, array_keys($options), array_values($options));
+
+        return sprintf('%s(%s)', $annotationClass, implode(', ', $formattedOptions));
+    }
+
+    private function quoteAnnotationValue($value)
+    {
+        if (is_bool($value)) {
+            return $value ? 'true' : 'false';
+        }
+
+        if (null === $value) {
+            return 'null';
+        }
+
+        if (is_int($value) || '0' === $value) {
+            return $value;
+        }
+
+        if (is_array($value)) {
+            throw new Exception('Invalid value: loop before quoting.');
+        }
+
+        return sprintf('"%s"', $value);
     }
 }