+++ CHANGELOG.md
@@ -9,12 +9,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - added @ApiSubresource annotation for tree entities
+- builder support for validating constraints and theire options
 
 ## [1.2.0] - 2020-01-28
 
 ### Changed
 
 - fixed some coding issues and continuous integration stuff
+- database table name conventions
 
 ## [1.1.0] - 2019-12-21